requirements.txt
gunicorn
django-heroku
django-bootstrap4
dj-database-url
psycopg2
social-auth-core
social-auth-app-django
selenium
pillow
django-crispy-forms